+++ packages/terra-date-picker/src/DateInput.jsx
@@ -578,10 +578,24 @@ const DatePickerInput = (props) => {
     }
     // decrement current valid date by 1 day, if not valid set date to yesterday instead
     if (event.key === '-' || event.key === '_') {
-      if (validDate) {
-        inputDate = DateUtil.decrementDateByDay(inputDate, DateUtil.ISO_EXTENDED_DATE_FORMAT);
-      } else {
-        inputDate = DateUtil.decrementDateByDay(DateUtil.getCurrentDate(), DateUtil.ISO_EXTENDED_DATE_FORMAT);
+      if (inputType === 0) {
+        if (validDate) {
+          inputDate = DateUtil.decrementDateByDay(inputDate, DateUtil.ISO_EXTENDED_DATE_FORMAT);
+        } else {
+          inputDate = DateUtil.decrementDateByDay(DateUtil.getCurrentDate(), DateUtil.ISO_EXTENDED_DATE_FORMAT);
+        }
+      } else if (inputType === 1) {
+        if (validDate) {
+          inputDate = DateUtil.decrementMonthByMonth(inputDate, DateUtil.ISO_EXTENDED_DATE_FORMAT);
+        } else {
+          inputDate = DateUtil.decrementMonthByMonth(DateUtil.getCurrentDate(), DateUtil.ISO_EXTENDED_DATE_FORMAT);
+        }
+      } else if (inputType === 2) {
+        if (validDate) {
+          inputDate = DateUtil.decrementYearByYear(inputDate, DateUtil.ISO_EXTENDED_DATE_FORMAT);
+        } else {
+          inputDate = DateUtil.decrementYearByYear(DateUtil.getCurrentDate(), DateUtil.ISO_EXTENDED_DATE_FORMAT);
+        }
       }
       formattedDate = DateUtil.strictFormatISODate(inputDate, momentDateFormat);
       if (onChange) {
@@ -593,10 +607,24 @@ const DatePickerInput = (props) => {
     }
     // increment current valid date by 1 day, if not valid date set date to tomorrow instead
     if (event.key === '=' || event.key === '+') {
-      if (validDate) {
-        inputDate = DateUtil.incrementDateByDay(inputDate, DateUtil.ISO_EXTENDED_DATE_FORMAT);
-      } else {
-        inputDate = DateUtil.incrementDateByDay(DateUtil.getCurrentDate(), DateUtil.ISO_EXTENDED_DATE_FORMAT);
+      if (inputType === 0) {
+        if (validDate) {
+          inputDate = DateUtil.incrementDateByDay(inputDate, DateUtil.ISO_EXTENDED_DATE_FORMAT);
+        } else {
+          inputDate = DateUtil.incrementDateByDay(DateUtil.getCurrentDate(), DateUtil.ISO_EXTENDED_DATE_FORMAT);
+        }
+      } else if (inputType === 1) {
+        if (validDate) {
+          inputDate = DateUtil.incrementMonthByMonth(inputDate, DateUtil.ISO_EXTENDED_DATE_FORMAT);
+        } else {
+          inputDate = DateUtil.incrementMonthByMonth(DateUtil.getCurrentDate(), DateUtil.ISO_EXTENDED_DATE_FORMAT);
+        }
+      } else if (inputType === 2) {
+        if (validDate) {
+          inputDate = DateUtil.incrementYearByYear(inputDate, DateUtil.ISO_EXTENDED_DATE_FORMAT);
+        } else {
+          inputDate = DateUtil.incrementYearByYear(DateUtil.getCurrentDate(), DateUtil.ISO_EXTENDED_DATE_FORMAT);
+        }
       }
       formattedDate = DateUtil.strictFormatISODate(inputDate, momentDateFormat);
       if (onChange) {
@@ -885,8 +913,8 @@ const DatePickerInput = (props) => {
         <div id={formatDescriptionId} className={cx('format-text')}>
           <VisuallyHiddenText
             aria-live={DateUtil.isMac() ? 'polite' : 'off'}
-            text={`${invalidEntry} 
-            ${intl.formatMessage({ id: 'Terra.datePicker.dateFormatLabel' })} 
+            text={`${invalidEntry}
+            ${intl.formatMessage({ id: 'Terra.datePicker.dateFormatLabel' })}
             ${format}. ${inputDate ? `${inputDate},` : ''} ${intl.formatMessage({ id: 'Terra.datePicker.hotKey' })} `}
           />
           <div aria-hidden="true">